文章发布

+++ src/main/java/com/sosd/repository/BlogDao.java
@@ -6,5 +6,5 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlogDao extends ElasticsearchRepository<Blog, Integer> {
+public interface BlogDao extends ElasticsearchRepository<Blog, Long> {
 }
